requirements.txt
Django==2.0.3
django-markdownx==2.0.22
pytz==2018.3
django-photologue==3.8.1
bootstrap-admin==0.3.9